requirements.txt
sentence-transformers
faiss-cpu
numpy
openai
fastapi
uvicorn
streamlit
pydantic
python-dotenv
requests
scikit-learn
